requirements.txt
tensorflow
matplotlib
numpy
pillow
tk
scikit-learn